+++ app/Http/Controllers/HandkerchiefController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ShowHandkerchiefRequest;
+use App\Http\Resources\HandkerchiefByIDResource;
 use App\Http\Resources\HandkerchiefHistoryResource;
 use App\Http\Resources\HandkerchiefResource;
 use App\Http\Resources\StoreHandkerchiefResource;
@@ -12,6 +13,7 @@ use App\Http\Requests\UpdateHandkerchiefRequest;
 use App\Models\HandkerchiefHistory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class HandkerchiefController extends Controller
@@ -24,6 +26,7 @@ class HandkerchiefController extends Controller
 
     public function index(Request $request)
     {
+        Gate::authorize('viewAny', Handkerchief::class);
         $handkerchief = Handkerchief::all();
         return HandkerchiefResource::collection($handkerchief);
     }
@@ -31,6 +34,7 @@ class HandkerchiefController extends Controller
 
     public function store(StoreHandkerchiefRequest $request)
     {
+        Gate::authorize('create', Handkerchief::class);
         $handkerchief = Handkerchief::create([
             "box_id" => $request->box_id,
             "name" => $request->name,
@@ -49,16 +53,44 @@ class HandkerchiefController extends Controller
      */
     public function show(Request $request, Handkerchief $handkerchief)
     {
-//        if ($request->filled('finished_products')) {
-//            $handkerchief->where('finished_products', $request->get('finished_products'));
-//        }
-        $history = $handkerchief->orderBy('finished_products');
-        return $this->reply($history);
+        if(!Gate::authorize('view', $handkerchief)){
+            return response()->json(["Sizda bu yerga kirish uchun ruxsat yo'q"], 403);
+        }else{
+//            if ($request->filled('storage_in')) {
+//               return HandkerchiefHistoryResource::collection($handkerchief->handkerchiefHistories()->where("storage_in",'=', true)->get());
+//            }
+//            if ($request->filled('sold_out')) {
+//                return HandkerchiefHistoryResource::collection($handkerchief->handkerchiefHistories()->where('sold_out', '=', true)->get());
+//            }
+
+            return new HandkerchiefByIDResource($handkerchief);
+//            $history = $handkerchief->orderBy('finished_products');
+////            dd($history);
+//            return $this->reply($history);
+
+        }
     }
+
+    public function viewHandkerchiefHistory(Request $request, Handkerchief $handkerchief)
+    {
+        if(!Gate::authorize('view', $handkerchief)){
+            return response()->json(["Sizda bu yerga kirish uchun ruxsat yo'q"], 403);
+        }else{
+            if ($request->filled('storage_in')) {
+               return HandkerchiefHistoryResource::collection($handkerchief->handkerchiefHistories()->where("storage_in",'=', true)->get());
+            }
+            if ($request->filled('sold_out')) {
+                return HandkerchiefHistoryResource::collection($handkerchief->handkerchiefHistories()->where('sold_out', '=', true)->get());
+            }
+        }
+    }
+
+
 
 
     public function update(UpdateHandkerchiefRequest $request, Handkerchief $handkerchief)
     {
+        Gate::authorize('update', Handkerchief::class);
             $handkerchief->name = $request->name;
             $handkerchief->sort_plane = $request->sort_plane;
             $handkerchief->save();
@@ -69,6 +101,7 @@ class HandkerchiefController extends Controller
      */
     public function destroy(Handkerchief $handkerchief)
     {
+        Gate::authorize('delete', Handkerchief::class);
         Storage::disk('public')->delete("$handkerchief->image_path");
         File::DeleteDirectory('storage/' . 'boxes/' . "$handkerchief->id");
         $handkerchief->delete();
